(session) compose session hash
Currently seeing a compiler bug when attempting to compile with raven.cr. Although I'm not convinced entirely that this is an issue with action-controller, it does seem to be considered good practice to stick with composition rather than inheritence for this sort of stuff (https://github.com/crystal-lang/crystal/issues/3238#issuecomment-244548787).

The approach taken was effectively the same as amberframework/amber#930 which fixed the same issue within Amber.